Remove real*fft related code
Currently it does not seem like it's possible to have a safe API for
real_fft and real_ifft operations.

Firstly, using these functions from outside the crate does not work at
all and causes an ICE:
https://github.com/rust-lang/rust/issues/101125

Secondly, rust does not allow for generic code that operates on arrays
without specifying the bounds. For this, rust needs to have a way of
expressing "unification" or "proving disjointness" for a set of
`impl`'s:
https://rust-lang.zulipchat.com/#narrow/stream/260443-project-const-generics/topic/Exhaustiveness.20Checking.20in.20Types

I've asked on reddit about this case:
https://www.reddit.com/r/rust/comments/wy9sk1/can_you_have_a_function_return_different_types/

And got response:
```
But Chalk may allow you to implement what you actually want (different
cases for odd and even numbers).
```